Update user interface and add README.md
The user interface of the AlphaFold Server CIF to PDB Converter is improved by adding a progress bar, and allowing for multiple file upload. Uses a markdown string with HTML edition activated for app description. Furthermore, a README.md file is added to provide users with all necessary instructions to use the application.